+++ R/az_group.R
@@ -1,6 +1,6 @@
 #' Group in Azure Active Directory
 #'
-#' Base class representing an AAD group.
+#' Class representing an AAD group.
 #'
 #' @docType class
 #' @section Fields:
@@ -18,6 +18,9 @@
 #' - `list_object_memberships()`: Return the IDs of all groups, administrative units and directory roles this group is a member of.
 #' - `list_members(type=c("user", "group", "application", "servicePrincipal"))`: Return a list of all members of this group. Specify the `type` argument to filter the result for specific object type(s).
 #' - `list_owners(type=c("user", "group", "application", "servicePrincipal"))`: Return a list of all owners of this group. Specify the `type` argument to filter the result for specific object type(s).
+#' - `get_sharepoint_site()`: Get the SharePoint site associated with this group, if it exists.
+#' - `list_drives()`: List the drives (shared document libraries) associated with this group.
+#' - `get_drive()`: Get the default document library for this group.
 #'
 #' @section Initialization:
 #' Creating new objects of this class should be done via the `create_group` and `get_group` methods of the [ms_graph] and [az_app] classes. Calling the `new()` method for this class only constructs the R object; it does not call the Microsoft Graph API to create the actual group.
@@ -26,7 +29,7 @@
 #' [ms_graph], [az_app], [az_user], [az_object]
 #'
 #' [Microsoft Graph overview](https://docs.microsoft.com/en-us/graph/overview),
-#' [REST API reference](https://docs.microsoft.com/en-us/graph/api/overview?view=graph-rest-beta)
+#' [REST API reference](https://docs.microsoft.com/en-us/graph/api/overview?view=graph-rest-1.0)
 #'
 #' @examples
 #' \dontrun{
@@ -50,24 +53,32 @@ public=list(
     initialize=function(token, tenant=NULL, properties=NULL)
     {
         self$type <- "group"
+        private$api_type <- "groups"
         super$initialize(token, tenant, properties)
     },
 
     list_members=function(type=c("user", "group", "application", "servicePrincipal"))
     {
         res <- private$get_paged_list(self$do_operation("members"))
-        private$init_list_objects(private$filter_list(res, type))
+        private$init_list_objects(res, type)
     },
 
     list_owners=function(type=c("user", "group", "application", "servicePrincipal"))
     {
         res <- private$get_paged_list(self$do_operation("owners"))
-        private$init_list_objects(private$filter_list(res, type))
+        private$init_list_objects(res, type)
     },
 
     print=function(...)
     {
-        cat("<Graph group '", self$properties$displayName, "'>\n", sep="")
+        group_type <- if("Unified" %in% self$properties$groupTypes)
+            "Microsoft 365"
+        else if(!self$properties$mailEnabled)
+            "Security"
+        else if(self$properties$securityEnabled)
+            "Mail-enabled security"
+        else "Distribution"
+        cat("<", group_type, " group '", self$properties$displayName, "'>\n", sep="")
         cat("  directory id:", self$properties$id, "\n")
         cat("  description:", self$properties$description, "\n")
         cat("---\n")