+++ R/az_object.R
@@ -1,6 +1,6 @@
 #' Azure Active Directory object
 #'
-#' Base class representing a directory object in Microsoft Graph.
+#' Base class representing an Azure Active Directory object in Microsoft Graph.
 #'
 #' @docType class
 #' @section Fields:
@@ -24,54 +24,13 @@
 #' [ms_graph], [az_app], [az_service_principal], [az_user], [az_group]
 #'
 #' [Microsoft Graph overview](https://docs.microsoft.com/en-us/graph/overview),
-#' [REST API reference](https://docs.microsoft.com/en-us/graph/api/overview?view=graph-rest-beta)
+#' [REST API reference](https://docs.microsoft.com/en-us/graph/api/overview?view=graph-rest-1.0)
 #'
-#' @format An R6 object of class `az_object`.
+#' @format An R6 object of class `az_object`, inheriting from `ms_object`.
 #' @export
-az_object <- R6::R6Class("az_object",
+az_object <- R6::R6Class("az_object", inherit=ms_object,
 
 public=list(
-
-    token=NULL,
-    tenant=NULL,
-    type=NULL,
-
-    # app data from server
-    properties=NULL,
-
-    initialize=function(token, tenant=NULL, properties=NULL)
-    {
-        self$token <- token
-        self$tenant <- tenant
-        self$properties <- properties
-    },
-
-    update=function(...)
-    {
-        self$do_operation(body=list(...), encode="json", http_verb="PATCH")
-        self$properties <- self$do_operation()
-        self
-    },
-
-    sync_fields=function()
-    {
-        self$properties <- self$do_operation()
-        invisible(self)
-    },
-
-    delete=function(confirm=TRUE)
-    {
-        if(confirm && interactive())
-        {
-            msg <- sprintf("Do you really want to delete the %s '%s'?",
-                           self$type, self$properties$displayName)
-            if(!get_confirmation(msg, FALSE))
-                return(invisible(NULL))
-        }
-
-        self$do_operation(http_verb="DELETE")
-        invisible(NULL)
-    },
 
     list_object_memberships=function()
     {
@@ -89,75 +48,13 @@ public=list(
         unlist(private$get_paged_list(lst))
     },
 
-    do_operation=function(op="", ...)
-    {
-        op <- construct_path(private$get_endpoint(), self$properties$id, op)
-        call_graph_endpoint(self$token, op, ...)
-    },
-
     print=function(...)
     {
-        cat("<Graph directory object '", self$properties$displayName, "'>\n", sep="")
+        cat("<Azure Active Directory object '", self$properties$displayName, "'>\n", sep="")
         cat("  directory id:", self$properties$id, "\n")
         cat("---\n")
         cat(format_public_methods(self))
         invisible(self)
     }
-),
-
-private=list(
-
-    get_paged_list=function(lst, next_link_name="@odata.nextLink", value_name="value")
-    {
-        res <- lst[[value_name]]
-        while(!is_empty(lst[[next_link_name]]))
-        {
-            lst <- call_graph_url(self$token, lst[[next_link_name]])
-            res <- c(res, lst[[value_name]])
-        }
-        res
-    },
-
-    filter_list=function(lst, type=c("user", "group", "application", "servicePrincipal", "device"))
-    {
-        type <- paste0("#microsoft.graph.", type)
-        keep <- vapply(lst, function(obj) obj$`@odata.type` %in% type, FUN.VALUE=logical(1))
-        lst[keep]
-    },
-
-    init_list_objects=function(lst)
-    {
-        lapply(lst, function(obj)
-        {
-            switch(obj$`@odata.type`,
-                "#microsoft.graph.user"=
-                    az_user$new(self$token, self$tenant, obj),
-                "#microsoft.graph.group"=
-                    az_group$new(self$token, self$tenant, obj),
-                "#microsoft.graph.application"=
-                    az_app$new(self$token, self$tenant, obj),
-                "#microsoft.graph.servicePrincipal"=
-                    az_service_principal$new(self$token, self$tenant, obj),
-                "#microsoft.graph.device"=
-                    az_device$new(self$token, self$tenant, obj),
-                {
-                    warning("Unknown directory object type ", obj$`@odata.type`)
-                    obj
-                }
-            )
-        })
-    },
-
-    get_endpoint=function()
-    {
-        switch(self$type,
-            "user"="users",
-            "group"="groups",
-            "application"="applications",
-            "service principal"="servicePrincipals",
-            "device"="devices",
-            stop("Unknown directory object type"))
-    }
 ))
-