+++ R/is.R
@@ -40,7 +40,48 @@ is_group <- function(object)
 
 #' @rdname info
 #' @export
-is_directory_object <- function(object)
+is_directory_role <- function(object)
+{
+    R6::is.R6(object) && inherits(object, "az_directory_role")
+}
+
+
+#' @rdname info
+#' @export
+is_sharepoint_site <- function(object)
+{
+    R6::is.R6(object) && inherits(object, "ms_site")
+}
+
+
+#' @rdname info
+#' @export
+is_drive <- function(object)
+{
+    R6::is.R6(object) && inherits(object, "ms_drive")
+}
+
+
+#' @rdname info
+#' @export
+is_sharepoint_list <- function(object)
+{
+    R6::is.R6(object) && inherits(object, "ms_sharepoint_list")
+}
+
+
+#' @rdname info
+#' @export
+is_aad_object <- function(object)
 {
     R6::is.R6(object) && inherits(object, "az_object")
 }
+
+
+#' @rdname info
+#' @export
+is_msgraph_object <- function(object)
+{
+    R6::is.R6(object) && inherits(object, "ms_object")
+}
+