+++ R/ms_graph.R
@@ -17,8 +17,6 @@
 #' - `create_group(name, email, ...)`: Creates a new group. Note that only security groups can be created via the Microsoft Graph API.
 #' - `get_group(group_id)`: Retrieves an existing group.
 #' - `delete_group(group_id, confirm=TRUE)`: Deletes a group.
-#' - `get_sharepoint_site(site_url, site_id)`: Retrieves a SharePoint site, either by URL or ID.
-#' - `get_drive(drive_id`): Retrieves a drive by ID. This can either be a user's OneDrive, or a shared document library associated with a group or SharePoint site. Note that it's probably easier to use the `get_drive` method for the associated user, group or site object.
 #' - `call_graph_endpoint(op="", ...)`: Calls the Microsoft Graph API using this object's token and tenant as authentication arguments. See [call_graph_endpoint].
 #'
 #' @section Authentication: