+++ R/zzz_class_directory.R
@@ -4,7 +4,7 @@
 #' @param R6_generator An R6 class generator corresponding to this Graph class.
 #' @param check_function A boolean function that checks if a list of properties is for an object of this class.
 #' @details
-#' As written, AzureGraph knows about a subset of all the object classes contained in Microsoft Graph. These are mostly the classes originating from Azure Active Directory and Microsoft Office: users, groups, app registrations, service principals, SharePoint sites and OneDrive libraries.
+#' As written, AzureGraph knows about a subset of all the object classes contained in Microsoft Graph. These are mostly the classes originating from Azure Active Directory: users, groups, app registrations, service principals and registered devices.
 #'
 #' You can extend AzureGraph by writing your own R6 class that inherits from `ms_object`. If so, you should also _register_ your class by calling `register_graph_class` and providing the generator object, along with a check function. The latter should accept a list of object properties (as obtained from the Graph REST API), and return TRUE/FALSE based on whether the object is of your class.
 #'